refactor: port utils.lazy_tools into its only use case, utils.pixelated_stem_tools
